requirements.txt
Django~=3.1.3
channels~=2.1.2

django-configurations~=2.2